+++ changelog.markdown
@@ -1,6 +1,7 @@
 # 1.2.0 Lexingtown
 
 - Replaced `marked` with `markdown-it`, improvements in compliance, parser speed, and code structure
+- Exposed `markdown-it` parser instance for extensibility
 
 # 1.1.0 IPO